Now takes into account all folders
Originally the python that created the structure was limited to Python2, due to changes to generators and the removal of next(). I didn't like that, so I rewrote how the json was being created. Not only does it work on Python 2 & 3, but it also takes into account all markdown - not just the markdown in the 'markdown' folder